Refactor to not depend on zustand
