+++ tierkreis_visualization/frontend/src/components/types.ts
@@ -2,6 +2,10 @@ export interface InfoProps {
   type: "Logs" | "Errors";
   content: string;
 }
+export interface Workflow {
+  id: string;
+  name: string;
+}
 
 export interface HandleProps {
   handles: string[];
