+++ tierkreis_visualization/frontend/src/components/workflow-sidebar.tsx
@@ -1,5 +1,3 @@
-import { useEffect } from "react";
-
 import {
   Sidebar,
   SidebarContent,
@@ -10,103 +8,16 @@ import {
   SidebarMenuButton,
   SidebarMenuItem,
 } from "@/components/ui/sidebar";
-import { Link, useParams } from "react-router";
+import { Link } from "react-router";
+import { Workflow } from "@/components/types";
 
-import { URL } from "@/data/constants";
-import useStore from "@/data/store";
-import { parseGraph } from "@/graph/parseGraph";
-import { useQuery, useQueryClient } from "@tanstack/react-query";
-interface Workflow {
-  id: string;
-  name: string;
-}
-
-function useWorkflows(url: string) {
-  return useQuery<Workflow[]>({
-    queryKey: ["workflows", url],
-    queryFn: async () => {
-      const response = await fetch(`${url}/all`);
-      if (!response.ok) {
-        throw new Error("Network response was not ok");
-      }
-      return response.json();
-    },
-    select: (data) =>
-      data.map(
-        (workflow): Workflow => ({
-          id: workflow.id,
-          name: workflow.name,
-        })
-      ),
-    staleTime: 1000 * 60 * 5, // Cache for 5 minutes
-  });
-}
-
-export function WorkflowSidebar() {
-  const { data: items = [] } = useWorkflows(URL);
-  const { workflowId = "" } = useParams();
-  const {
-    setWorkflowId,
-    setEdges,
-    setNodes,
-    clearOldEdges,
-    recalculateNodePositions,
-    tryFromStorage,
-  } = useStore();
-  const { clear } = useStore.temporal.getState();
-  const queryClient = useQueryClient();
-
-  const { data: graph } = useQuery({
-    queryKey: ["workflowGraph", workflowId],
-    queryFn: async () => {
-      console.log("Fetching graph for workflowId:", workflowId);
-      const response = await fetch(`${URL}/${workflowId}/nodes/-`, {
-        method: "GET",
-        headers: { Accept: "application/json" },
-      });
-      if (!response.ok) {
-        throw new Error("Network response was not ok");
-      }
-      return response.json();
-    },
-    enabled: !!workflowId,
-    select: (data) => parseGraph(data, workflowId),
-    // initialData
-  });
-  useEffect(() => {
-    setNodes([], false);
-    setEdges([]);
-    if (!tryFromStorage(workflowId) && graph) {
-      setNodes(graph.nodes, false);
-      setEdges(graph.edges);
-      recalculateNodePositions();
-    }
-  }, [
-    graph,
-    workflowId,
-    recalculateNodePositions,
-    setEdges,
-    setNodes,
-    tryFromStorage,
-  ]);
-  useEffect(() => {
-    if (workflowId) {
-      setWorkflowId(workflowId);
-      const ws = new WebSocket(`${URL}/${workflowId}/nodes/-`);
-      ws.onmessage = (event) => {
-        const graph = parseGraph(JSON.parse(event.data), workflowId);
-        queryClient.setQueryData(["workflowGraph", workflowId], graph);
-        setNodes(graph.nodes, false);
-        setEdges(graph.edges);
-      };
-      return () => {
-        if (ws.readyState == WebSocket.OPEN) {
-          ws.close();
-        }
-      };
-    }
-  }, [workflowId, queryClient, setWorkflowId, setEdges, setNodes]);
-
+export function WorkflowSidebar({
+  workflows,
+  workflowId,
+}: {
+  workflows: Workflow[];
+  workflowId: string;
+}) {
   return (
     <Sidebar>
       <SidebarContent>
@@ -114,18 +25,14 @@ export function WorkflowSidebar() {
           <SidebarGroupLabel>Workflows </SidebarGroupLabel>
           <SidebarGroupContent>
             <SidebarMenu>
-              {items.map((item) => (
-                <SidebarMenuItem key={item.id}>
+              {workflows.map((workflow) => (
+                <SidebarMenuItem key={workflow.id}>
                   <SidebarMenuButton
                     asChild
-                    isActive={workflowId === item.id}
-                    onClick={() => {
-                      clearOldEdges(); // not part of the tracked state
-                      clear();
-                    }}
+                    isActive={workflowId === workflow.id}
                   >
-                    <Link to={`/${item.id}`}>
-                      <span>{item.name}</span>
+                    <Link to={`/${workflow.id}`}>
+                      <span>{workflow.name}</span>
                     </Link>
                   </SidebarMenuButton>
                 </SidebarMenuItem>
